# auth-service/docs/swagger.yaml
basePath: /
definitions:
  handler.APIErrorResponse:
    properties:
      code:
        example: AUTH_401_01
        type: string
      details:
        items:
          $ref: '#/definitions/handler.ErrorDetail'
        type: array
      error:
        example: UNAUTHORIZED
        type: string
      message:
        example: Email atau kata sandi tidak cocok.
        type: string
      success:
        example: false
        type: boolean
      timestamp:
        example: "2026-09-25T09:30:00Z"
        type: string
    type: object
  handler.APIResponse:
    properties:
      data: {}
      message:
        example: Operasi berhasil
        type: string
      success:
        example: true
        type: boolean
    type: object
  handler.AuthContractData:
    properties:
      accessToken:
        example: eyJhbGciOi...
        type: string
      expiresIn:
        example: 900
        type: integer
      refreshToken:
        example: rfk-94a28f73...
        type: string
      tokenType:
        example: Bearer
        type: string
      user:
        $ref: '#/definitions/handler.UserContractResponse'
    type: object
  handler.BaseResponseData:
    properties:
      error:
        example: error description
        type: string
      message:
        example: operation successful
        type: string
      status:
        example: healthy
        type: string
      success:
        example: true
        type: boolean
    type: object
  handler.ErrorDetail:
    properties:
      field:
        example: password
        type: string
      message:
        example: Kata sandi salah
        type: string
    type: object
  handler.RefreshContractData:
    properties:
      accessToken:
        example: eyJhbGciOi...
        type: string
      expiresIn:
        example: 900
        type: integer
      refreshToken:
        example: rfk-new-rotated...
        type: string
      tokenType:
        example: Bearer
        type: string
    type: object
  handler.RefreshTokenResponseData:
    properties:
      access_token:
        example: eyJhbGciOi...
        type: string
      refresh_token:
        example: d0a5f9...
        type: string
      success:
        example: true
        type: boolean
    type: object
  handler.UserContractResponse:
    properties:
      avatar:
        type: string
      createdAt:
        type: string
      devices:
        type: integer
      email:
        type: string
      id:
        type: string
      memberSince:
        type: string
      name:
        type: string
      role:
        type: string
      tier:
        type: string
      watchHours:
        type: number
    type: object
  handler.changePasswordReq:
    properties:
      currentPassword:
        example: LiveEuy#2026
        type: string
      newPassword:
        example: LiveEuyUltra#2027
        minLength: 6
        type: string
    required:
    - currentPassword
    - newPassword
    type: object
  handler.demoLoginReq:
    properties:
      persona:
        example: hafiz
        type: string
    required:
    - persona
    type: object
  handler.forgotPasswordReq:
    properties:
      email:
        example: arga@example.com
        type: string
    required:
    - email
    type: object
  handler.loginReq:
    properties:
      email:
        example: hafiz@liveeuy.id
        type: string
      password:
        example: LiveEuy#2026
        type: string
      rememberMe:
        example: true
        type: boolean
    required:
    - email
    - password
    type: object
  handler.logoutReq:
    properties:
      refresh_token:
        example: rfk-94a28f73b610c41d99e52e
        type: string
      refreshToken:
        example: rfk-94a28f73b610c41d99e52e
        type: string
    type: object
  handler.profileUpdateReq:
    properties:
      avatar:
        example: https://images.unsplash.com/photo-custom.jpg
        type: string
      name:
        example: Hafiz Muhammad Al Ikhsan
        type: string
    type: object
  handler.refreshTokenReq:
    properties:
      refresh_token:
        example: rfk-94a28f73b610c41d99e52e
        type: string
      refreshToken:
        example: rfk-94a28f73b610c41d99e52e
        type: string
    type: object
  handler.registerReq:
    properties:
      email:
        example: arga@example.com
        type: string
      name:
        example: Arga Pratama
        type: string
      password:
        example: PasswordSuper#2026
        minLength: 6
        type: string
      tier:
        example: VIP Standard
        type: string
    required:
    - email
    - name
    - password
    type: object
  handler.resetPasswordReq:
    properties:
      newPassword:
        example: NewSecurePassword#2026
        minLength: 6
        type: string
      token:
        example: rst-tok-84729104857201
        type: string
    required:
    - newPassword
    - token
    type: object
host: localhost:8080
info:
  contact:
    email: support@liveeuy.com
    name: API Support
  description: Authentication & User Management Service for LiveEuy platform.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Auth Service API
  version: "1.0"
paths:
  /.well-known/jwks.json:
    get:
      description: Exposes RSA public keys in RFC 7517 format for downstream microservices
      produces:
      - application/json
      responses:
        "200":
          description: JSON Web Key Set
          schema:
            additionalProperties: true
            type: object
      summary: JSON Web Key Set (Public Keys)
      tags:
      - System
  /api/health:
    get:
      description: Service health status
      produces:
      - application/json
      responses:
        "200":
          description: Healthy
          schema:
            $ref: '#/definitions/handler.BaseResponseData'
      summary: Health check
      tags:
      - System
  /api/ping:
    get:
      description: Ping test endpoint
      produces:
      - application/json
      responses:
        "200":
          description: Pong
          schema:
            $ref: '#/definitions/handler.BaseResponseData'
      summary: Ping check
      tags:
      - System
  /api/v1/auth/change-password:
    put:
      consumes:
      - application/json
      description: Change password with validation against current password
      parameters:
      - description: Change Password Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.changePasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: Kata sandi berhasil diubah
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIErrorResponse'
        "401":
          description: Kata sandi saat ini salah
          schema:
            $ref: '#/definitions/handler.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Change user password
      tags:
      - Authentication
  /api/v1/auth/demo-login:
    post:
      consumes:
      - application/json
      description: Fast login for development, testing, and UI persona testing (hafiz
        / budi)
      parameters:
      - description: Demo Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.demoLoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil masuk via Demo Account
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.AuthContractData'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.APIErrorResponse'
      summary: Demo persona login
      tags:
      - Authentication
  /api/v1/auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Request password reset email
      parameters:
      - description: Forgot Password Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.forgotPasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: Email pemulihan terkirim
          schema:
            $ref: '#/definitions/handler.APIResponse'
      summary: Forgot password
      tags:
      - Authentication
  /api/v1/auth/google/callback:
    get:
      description: Handles OAuth callback from Google, authenticates the user and
        returns access and refresh tokens
      parameters:
      - description: OAuth State parameter
        in: query
        name: state
        required: true
        type: string
      - description: OAuth Authorization Code
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Google login successful
          schema:
            $ref: '#/definitions/handler.RefreshTokenResponseData'
        "400":
          description: Missing authorization code
          schema:
            $ref: '#/definitions/handler.BaseResponseData'
        "401":
          description: CSRF state mismatch or invalid session
          schema:
            $ref: '#/definitions/handler.BaseResponseData'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.BaseResponseData'
      summary: Google OAuth callback
      tags:
      - OAuth
  /api/v1/auth/google/login:
    get:
      description: Redirects the user to Google OAuth consent page
      responses:
        "307":
          description: Redirect to Google OAuth consent URL
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.BaseResponseData'
      summary: Initiate Google OAuth login
      tags:
      - OAuth
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user using email and password
      parameters:
      - description: Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.loginReq'
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil masuk ke LiveEuy
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.AuthContractData'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIErrorResponse'
        "401":
          description: Email atau kata sandi tidak cocok
          schema:
            $ref: '#/definitions/handler.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.APIErrorResponse'
      summary: User login
      tags:
      - Authentication
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidate the specified refresh token session and clear auth cookie
      parameters:
      - description: Logout Request (Optional if using Cookie)
        in: body
        name: request
        schema:
          $ref: '#/definitions/handler.logoutReq'
      produces:
      - application/json
      responses:
        "200":
          description: Sesi berhasil diakhiri
          schema:
            $ref: '#/definitions/handler.APIResponse'
      summary: User logout (single device)
      tags:
      - Authentication
  /api/v1/auth/logout-all:
    post:
      description: Invalidate all active refresh tokens for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Sesi di seluruh perangkat berhasil diakhiri
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout all devices
      tags:
      - Authentication
  /api/v1/auth/me:
    get:
      description: Retrieve current authenticated user profile
      produces:
      - application/json
      responses:
        "200":
          description: Profil berhasil dimuat
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.UserContractResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handler.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - Authentication
  /api/v1/auth/profile:
    put:
      consumes:
      - application/json
      description: Update authenticated user's name and avatar
      parameters:
      - description: Profile Update Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.profileUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: Profil berhasil diperbarui
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.UserContractResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - Authentication
  /api/v1/auth/public-key.pem:
    get:
      description: Exposes RSA public key in standard PEM format
      produces:
      - text/plain
      responses:
        "200":
          description: RSA Public Key PEM
          schema:
            type: string
      summary: RSA Public Key in PEM format
      tags:
      - System
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Rotate refresh token and generate new access token (from Cookie
        or JSON body)
      parameters:
      - description: Refresh Token Request (Optional if using Cookie)
        in: body
        name: request
        schema:
          $ref: '#/definitions/handler.refreshTokenReq'
      produces:
      - application/json
      responses:
        "200":
          description: Token akses berhasil diperbarui
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.RefreshContractData'
              type: object
        "401":
          description: Refresh token tidak valid atau telah dicabut
          schema:
            $ref: '#/definitions/handler.APIErrorResponse'
      summary: Refresh access token
      tags:
      - Authentication
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with name, email, password, and tier
      parameters:
      - description: Register Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.registerReq'
      produces:
      - application/json
      responses:
        "201":
          description: Pendaftaran akun berhasil
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.AuthContractData'
              type: object
        "400":
          description: Bad Request / Validation Error
          schema:
            $ref: '#/definitions/handler.APIErrorResponse'
        "409":
          description: Email sudah terdaftar di sistem
          schema:
            $ref: '#/definitions/handler.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.APIErrorResponse'
      summary: Register a new user
      tags:
      - Authentication
  /api/v1/auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset password using received token
      parameters:
      - description: Reset Password Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.resetPasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: Kata sandi berhasil direset
          schema:
            $ref: '#/definitions/handler.APIResponse'
      summary: Reset password
      tags:
      - Authentication
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"